% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsf.R
\name{default_BSUB_config}
\alias{default_BSUB_config}
\title{Deafult BSUB setting to submit array jobs to LSF clusters.}
\usage{
default_BSUB_config()
}
\value{
A named list of BSUB settings.
}
\description{
Deafult BSUB setting to submit array jobs to 
LSF clusters. This is handy to obtain a default input, and
implement custom modificiations. These parameters are to be
used by function \code{run_lsf}.
}
\examples{
custom_BSUB = default_BSUB_config()
custom_BSUB$`-J` = "Myjob"
custom_BSUB$`-P` = "my_project"
custom_BSUB$`-q` = "my_queue"
}
\seealso{
run_lsf
}
