% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_PBSpro_config.R
\name{default_PBSpro_config}
\alias{default_PBSpro_config}
\title{Deafult QSUB setting to submit array jobs to PBSpro clusters.}
\usage{
default_PBSpro_config(
  project = "EASYPAR_Project",
  queue = "thin",
  walltime = "3:00:00",
  n_and_cpus = c(1, 16),
  jobID = "EASYPAR_Runner",
  output = "output.^array_index^.log",
  error = "error.^array_index^.err",
  J = ""
)
}
\value{
A named list of QSUB settings.
}
\description{
Deafult QSUB setting to submit array jobs to
PBSpro clusters. This is handy to obtain a default input, and
implement custom modificiations. These parameters are to be
used by function \code{run_slurm}.
}
\examples{
custom_QSUB = default_PBSpro_config(time = '12:00')
custom_QSUB$`-nodes` = 10
custom_QSUB$`--partition` = "my_queue"
}
\seealso{
run_lsf
}
