% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_QSUB_config.R
\name{default_QSUB_config}
\alias{default_QSUB_config}
\title{Deafult QSUB setting to submit array jobs to PBSpro clusters.}
\usage{
default_QSUB_config(
  project = "EASYPAR_Project",
  queue = "thin",
  walltime = "3:00:00",
  n_and_cpus = c(1, 16),
  jobID = "EASYPAR_Runner",
  output = "output.^array_index^.log",
  error = "error.^array_index^.err",
  J = ""
)
}
\value{
A named list of QSUB settings.
}
\description{
Deafult QSUB setting to submit array jobs to
PBSpro clusters. This is handy to obtain a default input, and
implement custom modificiations. These parameters are to be
used by function \code{run_PBSpro}.
}
\examples{
custom_QSUB = default_QSUB_config(walltime = '12:00:00')
custom_QSUB$`-l nodes=:ppn=` = c(1,8)
custom_QSUB$`-q` = "my_queue"
}
\seealso{
run_lsf
}
