% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterErrors.R
\name{filterErrors}
\alias{filterErrors}
\title{Filter out the tasks with errors.}
\usage{
filterErrors(R)
}
\arguments{
\item{R}{The list of outputs from \code{run}}
}
\value{
The list of tasks without errors
}
\description{
Filter out the tasks with errors.
}
\examples{
TODO
}
