% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a function with easypar}
\usage{
run(FUN, PARAMS, packages = NULL, export = NULL, cores.ratio = 0.8,
  parallel = TRUE, silent = FALSE, outfile = "", cache = NULL,
  progress_bar = TRUE)
}
\arguments{
\item{FUN}{R function to run}

\item{PARAMS}{List of parameters, each entry needs to match the arguments of \code{FUN}}

\item{packages}{Packages to load in each parallel thread}

\item{export}{Variables to export to each parallel thread}

\item{cores.ratio}{Ratio of the available cores that are used for a parallel run, default 80\%.}

\item{parallel}{Boolean value, TRUE for a parallel execution, and FALSE for a standard for loop.}

\item{silent}{Silent output from easypar.}

\item{outfile}{Output file for the parallel thread, default \code{""} is console, \code{NA} is the default.}

\item{cache}{Cache is used during computation to dump results to a template RDS file.}
}
\value{

}
\description{
Run a task on a list of inputs, making easy to parallelize it or run it in
a sequential session. If the glonal option \code{easypar.parallel} is set, the parameter
\code{parallel} is overwritten. This allows to easily switch between debug and non-debug
execution modes for this task.
}
