% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{filterErrors}
\alias{filterErrors}
\title{Remove the tasks with errors from the output}
\usage{
filterErrors(R)
}
\arguments{
\item{R}{The list of outputs from \code{run}}
}
\value{
The list of tasks without errors
}
\description{
Remove the tasks with errors from the output
}
