% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterErrors.R
\name{filterErrors}
\alias{filterErrors}
\title{Filter out the tasks with errors.}
\usage{
filterErrors(R)
}
\arguments{
\item{R}{The list of outputs from \code{run}}
}
\value{
The list of tasks without errors
}
\description{
Filter out the tasks with errors.
}
\examples{
# Errors can be intercepted. Consider a function
# that can generate some error. The run will not plot and
# the computation will run anyway.

options(easypar.parallel = FALSE)

results = run(
  FUN = function(x) {
    if(runif(1) > .5) stop("Some error")
    x
  },
  PARAMS = params,
  silent = TRUE
)

# Getter that can filter them out

filterErrors(results)
}
