% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numErrors.R
\name{numErrors}
\alias{numErrors}
\title{Return the number of tasks with errors}
\usage{
numErrors(R)
}
\arguments{
\item{R}{The list of outputs from \code{run}}
}
\value{
The number of tasks with errors
}
\description{
Return the number of tasks with errors
}
\examples{
# Errors can be intercepted. Consider a function
# that can generate some error. The run will not plot and
# the computation will run anyway.

options(easypar.parallel = FALSE)

results = run(
  FUN = function(x) {
    if(runif(1) > .5) stop("Some error")
    x
  },
  PARAMS = params,
  silent = TRUE
)

# Getters that can return the number of errors

numErrors(results)
}
