% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numErrors.R
\name{numErrors}
\alias{numErrors}
\title{Return the number of tasks with errors}
\usage{
numErrors(R)
}
\arguments{
\item{R}{The list of outputs from \code{run}}
}
\value{
The number of tasks with errors
}
\description{
Return the number of tasks with errors
}
\examples{
TODO
}
